refactor: modify package structures
